typescript(vx-demo): Rewrite Pack demo in TypeScript
