+++ packages/vx-demo/src/pages/Pack.tsx
@@ -1,13 +1,20 @@
 import React from 'react';
+import Show from '../components/Show';
+import Pack from '../components/tiles/Pack';
+
+export default () => {
+  return (
+    <Show component={Pack} title="Pack">
+      {`import React from 'react';
 import { Group } from '@vx/group';
 import { Pack } from '@vx/hierarchy';
 import { hierarchy } from 'd3-hierarchy';
 import { scaleQuantize } from '@vx/scale';
 import { exoplanets as data } from '@vx/mock-data';
 
-const extent = (allData, value = d => d) => [
-  Math.min(...allData.map(value)),
-  Math.max(...allData.map(value)),
+const extent = (data, value = d => d) => [
+  Math.min(...data.map(value)),
+  Math.max(...data.map(value))
 ];
 
 const exoplanets = data.filter(d => d.distance === 0);
@@ -16,7 +23,7 @@ const pack = { children: [{ children: planets }].concat(exoplanets) };
 
 const colorScale = scaleQuantize({
   domain: extent(data, d => d.radius),
-  range: ['#ffe108', '#ffc10e', '#fd6d6f', '#855af2', '#11d2f9', '#49f4e7'],
+  range: ['#ffe108', '#ffc10e', '#fd6d6f', '#855af2', '#11d2f9', '#49f4e7']
 });
 
 export default ({
@@ -26,17 +33,14 @@ export default ({
     top: 10,
     left: 30,
     right: 40,
-    bottom: 80,
-  },
+    bottom: 80
+  }
 }) => {
-  if (width < 10) return null;
-
-  const root = hierarchy(pack)
+  const data = hierarchy(pack)
     .sum(d => d.radius * d.radius)
     .sort((a, b) => {
       return (
         !a.children - !b.children ||
-        // eslint-disable-next-line no-restricted-globals
         isNaN(a.data.distance) - isNaN(b.data.distance) ||
         a.data.distance - b.data.distance
       );
@@ -45,15 +49,15 @@ export default ({
   return (
     <svg width={width} height={height}>
       <rect width={width} height={height} rx={14} fill="#ffffff" />
-      <Pack root={root} size={[width * 2, height * 2]}>
-        {packData => {
-          const circles = packData.descendants().slice(2);
+      <Pack root={data} size={[width * 2, height * 2]}>
+        {pack => {
+          const circles = pack.descendants().slice(2);
           return (
             <Group top={-height - margin.bottom} left={-width / 2}>
               {circles.map((circle, i) => {
                 return (
                   <circle
-                    key={`cir-${i}`}
+                    key={\`cir-\${i}\`}
                     r={circle.r}
                     cx={circle.x}
                     cy={circle.y}
@@ -66,5 +70,9 @@ export default ({
         }}
       </Pack>
     </svg>
+  );
+};
+`}
+    </Show>
   );
 };
